ref: delete directory file_handler, and move file_handler.py to utils directory
